Fix timer sync issue when continuing a TE.
When there was a running time entry and the user continued another one
during that, the timer failed to pickup on the new time entry as it
expected that the IsRunning of the other entries would be set to false
before the IsRunning is set to true on the new one. In reality there are
two time entries running for a brief moment before the old one is
stopped.