Manually commit data on stop broadcast.
In case the Android system kills our process right after returning from
the broadcast receiver function we need to be sure that the data is
actually commited to the database. If it isn't the user gets the
perception that the time entry was stopped, but in the database it is
actually still running (and sync did also not happen).