Fix broken mapping between data and push tasks.
When we started removing remote rejected items from the graph before
pushing changes to the server we broke the assumption that tasks and
dirtyDataObjects would correspond one-to-one with eachother.

When the user had some items rejected by the server, we would not send
them to the server, but when we got the push results back, we would
operate on the remote rejected items, thus causing the valid ones to be
pushed twice (or possibly more times).

Closes #330.